Attempted to change the literal implementation to keep the lexical representation of Literals, and to use this lexical representation for equality checks.
This would mirror the usage in Jena.

Related to the dicussion of issue 211.

However, I ran into some problems, we already try very hard to do "the sensible thing" when comparing literals, with other literals, as well as with python objects.

i.e. currently all of these are true: Literal(1)<2, Literal("1")<2, Literal(1)<Literal(2.0)

i.e. not only do we ignore lexical differences, we also ignore data-types

We *could* keep this behaviour, and change equals to be strictly lexical.
This leads to the weird behaviour that:
>>> l1=Literal("01", datatype=XSD.integer)
>>> l2=Literal("1", datatype=XSD.integer)
>>> l1>l2 or l1<l2 or l1==l2
False

Though... which is not FORBIDDEN by python, but I'm pretty sure it's discouraged :)

I'll try the other way around :)