Another attempt at fixing literal equality, hash and comparison handling. Here lexical representations and values are kept separately, but all comparisons happen in value space.
This is keeping the existing rdflib behaviour for literal > < comparisons, here everything is compared based on the value. This means we can do some neat things:

* Compare literals with python objects, i.e. Literal(1)<2
* Compare literals without regard to datatype: Literal(1)<Literal(2.0)

BUT

This means we more completely ignore http://www.w3.org/TR/rdf-concepts/#section-Literal-Equality
Which is quite clear that data-type and lang must match for equality.

Now, to make __eq__, __gt__ and __hash__ consistent this must be true:

>>> Literal('1', datatype=XSD.integer) == Literal('1.0', datatype=XSD.float)
True

Which is probably not what we want.

I belive the neat comparisons will have to go...